Move Lockable to protocol to its own file
